Code cleanup and reduce unrelated code changes
