+++ money/__init__.py
@@ -1,4 +1,3 @@
-from money.exceptions import IncorrectMoneyInputError
-from money.money import CURRENCY, Currency, Money
+from money.money import CURRENCY, Currency, IncorrectMoneyInputError, Money
 
 __all__ = ["Money", "Currency", "CURRENCY", "IncorrectMoneyInputError"]
