+++ money/contrib/django/models/fields.py
@@ -1,192 +1,34 @@
-from decimal import Decimal
-from typing import TYPE_CHECKING, Any
-
 from django.db import models
-from django.db.models import Lookup
 from django.utils.translation import gettext_lazy
 
 from money.contrib.django import forms
-from money.money import Currency, Money
+from money.contrib.django.models.lookups import (
+    MoneyExactLookup,
+    MoneyGteLookup,
+    MoneyGtLookup,
+    MoneyLteLookup,
+    MoneyLtLookup,
+)
+from money.contrib.django.models.proxy import MoneyFieldProxy
+from money.contrib.django.models.utils import (
+    currency_field_db_column,
+    currency_field_name,
+)
+from money.money import Money
 
-if TYPE_CHECKING:
-    from django.db.backends.base.base import BaseDatabaseWrapper
-    from django.db.models.expressions import Col
-    from django.db.models.sql.compiler import SQLCompiler
-
-
-__all__ = ("MoneyField", "currency_field_name", "NotSupportedLookup")
-
-
-def currency_field_name(name: str) -> str:
-    return "%s_currency" % name
-
-
-def currency_field_db_column(db_column):
-    return None if db_column is None else "%s_currency" % db_column
-
+__all__ = ("MoneyField", "NotSupportedLookup")
 
 SUPPORTED_LOOKUPS = ("exact", "lt", "gt", "lte", "gte", "isnull")
 
 
 class NotSupportedLookup(TypeError):
     def __init__(self, lookup):
-        super().__init__()
+        super(NotSupportedLookup, self).__init__()
+
         self.lookup = lookup
 
     def __str__(self):
         return "Lookup '%s' is not supported for MoneyField" % self.lookup
-
-
-class MoneyCurrencyLookupMixin:
-    """
-    Mixin for lookups that need to check currency when Money objects are used.
-
-    This mixin expects to be mixed with django.db.models.Lookup and uses
-    the following attributes/methods from that class:
-    - process_lhs(): Process left-hand side of the lookup
-    - process_rhs(): Process right-hand side of the lookup
-    - lhs: Left-hand side expression (Col)
-    - rhs: Right-hand side value
-    - operator: SQL operator string
-    """
-
-    if TYPE_CHECKING:
-        # Type stubs for attributes/methods provided by Lookup base class
-        operator: str
-        rhs: Any
-        lhs: "Col"
-
-        def process_lhs(
-            self, compiler: "SQLCompiler", connection: "BaseDatabaseWrapper"
-        ) -> tuple[str, list[Any]]: ...
-
-        def process_rhs(
-            self, compiler: "SQLCompiler", connection: "BaseDatabaseWrapper"
-        ) -> tuple[str, list[Any]]: ...
-
-    def as_sql(
-        self, compiler: "SQLCompiler", connection: "BaseDatabaseWrapper"
-    ) -> tuple[str, list[Any]]:
-        lhs, lhs_params = self.process_lhs(compiler, connection)
-        rhs, rhs_params = self.process_rhs(compiler, connection)
-
-        # Check if the original rhs value (before processing) is a Money object
-        if isinstance(self.rhs, Money):
-            # Get the currency field column name
-            currency_field = currency_field_name(self.lhs.field.name)
-            currency_column = self.lhs.field.model._meta.get_field(
-                currency_field
-            ).column
-
-            # Build SQL that checks both amount and currency
-            # Get the table alias from lhs
-            table_name = self.lhs.alias
-            currency_lhs_sql = f"{compiler.quote_name_unless_alias(table_name)}.{compiler.quote_name_unless_alias(currency_column)}"
-
-            # Construct the full condition
-            amount_condition = f"{lhs} {self.operator} {rhs}"
-            currency_condition = f"{currency_lhs_sql} = %s"
-
-            sql = f"({amount_condition} AND {currency_condition})"
-            params = lhs_params + [self.rhs.amount, self.rhs.currency.code]
-
-            return sql, params
-        else:
-            # Normal lookup without currency check
-            return f"{lhs} {self.operator} {rhs}", lhs_params + rhs_params
-
-
-class MoneyExactLookup(MoneyCurrencyLookupMixin, Lookup):
-    lookup_name = "exact"
-    operator = "="
-
-
-class MoneyLtLookup(MoneyCurrencyLookupMixin, Lookup):
-    lookup_name = "lt"
-    operator = "<"
-
-
-class MoneyLteLookup(MoneyCurrencyLookupMixin, Lookup):
-    lookup_name = "lte"
-    operator = "<="
-
-
-class MoneyGtLookup(MoneyCurrencyLookupMixin, Lookup):
-    lookup_name = "gt"
-    operator = ">"
-
-
-class MoneyGteLookup(MoneyCurrencyLookupMixin, Lookup):
-    lookup_name = "gte"
-    operator = ">="
-
-
-class MoneyFieldProxy(object):
-    """
-    An equivalent to Django's default attribute descriptor class SubfieldBase
-    (normally enabled via `__metaclass__ = models.SubfieldBase` on the custom
-    Field class).
-
-    Instead of calling to_python() on our MoneyField class as SubfieldBase
-    does, it stores the two different parts separately, and updates them
-    whenever something is assigned. If the attribute is read, it builds the
-    instance "on-demand" with the current data.
-
-    See: http://blog.elsdoerfer.name/2008/01/08/fuzzydates-or-one-django-model-field-multiple-database-columns/
-    """
-
-    def __init__(self, field: "MoneyField"):
-        self.field: "MoneyField" = field
-        self.amount_field_name: str = field.name
-        self.currency_field_name: str = field.currency_field_name
-
-    def _get_values(self, obj: models.Model) -> tuple[Decimal | None, str | None]:
-        return (
-            obj.__dict__.get(self.amount_field_name, None),
-            obj.__dict__.get(self.currency_field_name, None),
-        )
-
-    def _set_values(
-        self,
-        obj: models.Model,
-        amount: Decimal | None,
-        currency: str | Currency | None,
-    ) -> None:
-        obj.__dict__[self.amount_field_name] = amount
-        obj.__dict__[self.currency_field_name] = currency
-
-    def __get__(self, obj: models.Model, *args):
-        if obj is None:
-            return self
-        amount, currency = self._get_values(obj)
-        if amount is None:
-            return None
-        return Money(amount, currency)
-
-    def __set__(self, obj: models.Model, value):
-        if value is None:  # Money(0) is False
-            self._set_values(obj, None, "")
-        elif isinstance(value, Money):
-            self._set_values(obj, value.amount, value.currency.code)
-        elif isinstance(value, Decimal):
-            _, currency = self._get_values(obj)  # use what is currently set
-            self._set_values(obj, value, currency)
-        else:
-            # It could be an int, or some other python native type
-            try:
-                amount = Decimal(str(value))
-                _, currency = self._get_values(obj)  # use what is currently set
-                self._set_values(obj, amount, currency)
-            except TypeError:
-                # Lastly, assume string type 'XXX 123' or something Money can
-                # handle.
-                try:
-                    _, currency = self._get_values(obj)  # use what is currently set
-                    m = Money.from_string(str(value))
-                    self._set_values(obj, m.amount, m.currency)
-                except TypeError:
-                    msg = 'Cannot assign "%s"' % type(value)
-                    raise TypeError(msg)
 
 
 class InfiniteDecimalField(models.DecimalField):
@@ -196,7 +38,7 @@ class InfiniteDecimalField(models.DecimalField):
         if "postgresql" in engine:
             return "numeric"
 
-        return super().db_type(connection=connection)
+        return super(InfiniteDecimalField, self).db_type(connection=connection)
 
     def get_db_prep_save(self, value, *args, **kwargs):
         """
@@ -232,6 +74,8 @@ class CurrencyField(models.CharField):
 
 class MoneyField(InfiniteDecimalField):
     description = gettext_lazy("An amount and type of currency")
+
+    add_currency_field: bool
     currency_field_name: str
 
     # Don't extend SubfieldBase since we need to have access to both fields when
@@ -243,7 +87,6 @@ class MoneyField(InfiniteDecimalField):
         default_currency = kwargs.pop("default_currency", "")
         default = kwargs.get("default", None)
         self.add_currency_field = not kwargs.pop("no_currency_field", False)
-
         self.blankable = kwargs.get("blank", False)
 
         if isinstance(default, Money):
@@ -252,10 +95,10 @@ class MoneyField(InfiniteDecimalField):
         else:
             self.default_currency = default_currency or ""  # use the kwarg passed in
 
-        super().__init__(*args, **kwargs)
+        super(MoneyField, self).__init__(*args, **kwargs)
 
     def deconstruct(self):
-        name, path, args, kwargs = super().deconstruct()
+        name, path, args, kwargs = super(MoneyField, self).deconstruct()
         kwargs["no_currency_field"] = True
         return name, path, args, kwargs
 
@@ -296,7 +139,7 @@ class MoneyField(InfiniteDecimalField):
             cls.add_to_class(self.currency_field_name, c_field)
 
         # Set ourselves up normally
-        super().contribute_to_class(cls, name)
+        super(MoneyField, self).contribute_to_class(cls, name)
 
         # As we are not using SubfieldBase, we need to set our proxy class here
         setattr(cls, self.name, MoneyFieldProxy(self))
@@ -312,7 +155,7 @@ class MoneyField(InfiniteDecimalField):
         if isinstance(value, Money):
             value = value.amount
 
-        return super().get_db_prep_save(value, *args, **kwargs)
+        return super(MoneyField, self).get_db_prep_save(value, *args, **kwargs)
 
     def get_db_prep_value(self, value, connection, prepared=False):
         """
@@ -334,7 +177,7 @@ class MoneyField(InfiniteDecimalField):
         if isinstance(self.default, Money):
             return self.default
         else:
-            return super().get_default()
+            return super(MoneyField, self).get_default()
 
     def value_to_string(self, obj):
         """
@@ -348,7 +191,7 @@ class MoneyField(InfiniteDecimalField):
     def formfield(self, **kwargs):
         defaults = {"form_class": forms.MoneyField}
         defaults.update(kwargs)
-        return super().formfield(**defaults)
+        return super(MoneyField, self).formfield(**defaults)
 
     @property
     def validators(self):
