+++ money/contrib/django/models/fields.pyi
@@ -1,13 +1,11 @@
-from typing import Any, Optional, TypeVar, Generic
+from decimal import Decimal
+from typing import Any, TypeVar
 
 from django.db import models
 from django.db.models import Combinable
-
-from money.money import Money, Currency
 from django.utils.functional import _StrOrPromise
-from decimal import Decimal
 
-def currency_field_name(name: str) -> str: ...
+from money.money import Currency, Money
 
 class NotSupportedLookup(TypeError): ...
 
@@ -16,19 +14,6 @@ class CurrencyField(models.CharField[str, str]):
     _pyi_private_get_type: str
 
 F = TypeVar("F", bound="MoneyField")
-
-class MoneyFieldProxy(Generic[F]):
-    field: F = ...
-    amount_field_name: str = ...
-    currency_field_name: str = ...
-
-    def __init__(self, field: F) -> None: ...
-    def _get_values(
-        self, obj: models.Model
-    ) -> tuple[Optional[Decimal], Optional[str]]: ...
-    def _set_values(
-        self, obj: models.Model, amount: Optional[Decimal], currency: Optional[str]
-    ) -> None: ...
 
 class MoneyField(models.DecimalField[Money, Money]):
     _pyi_private_set_type: Money | Decimal | int  # type: ignore[assignment]
@@ -41,7 +26,7 @@ class MoneyField(models.DecimalField[Money, Money]):
     decimal_places: int
 
     default_currency: str | Currency
-    no_currency_field: bool
+    add_currency_field: bool
 
     def __init__(
         self,
