+++ money/tests/test_money.py
@@ -2,8 +2,7 @@ from decimal import Decimal
 
 import pytest
 
-from money.exceptions import IncorrectMoneyInputError
-from money.money import CURRENCY, Currency, Money
+from money.money import CURRENCY, Currency, IncorrectMoneyInputError, Money
 
 
 def test_string_parse() -> None:
