+++ money/tests/test_money_operations.py
@@ -3,8 +3,7 @@ from typing import Callable
 
 import pytest
 
-from money.exceptions import CurrencyMismatchException, InvalidOperationException
-from money.money import Money
+from money.money import CurrencyMismatchException, InvalidOperationException, Money
 
 MONEY_STRINGS: list[tuple[Money, str]] = [
     # Undefined currency:
